pyproject.toml
[project]
name = "shoping-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "ersa rani", email = "ersarani955@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.6.3",
    "google-generativeai>=0.8.5",
    "openai-agents>=0.2.3",
]

[project.scripts]
shoping-agent = "shoping_agent:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
